update the ernie base model
