+++ PaddleNLP/lexical_analysis/run_ernie.sh
@@ -1,50 +1,53 @@
-set -eux
-
-export FLAGS_eager_delete_tensor_gb=0.0
-export FLAGS_fast_eager_deletion_mode=1
+#set -eux
+#export FLAGS_fraction_of_gpu_memory_to_use=0.02
+export FLAGS_eager_delete_tensor_gb=1.0
+#export FLAGS_fast_eager_deletion_mode=1
 export FLAGS_sync_nccl_allreduce=1
-export FLAGS_selected_gpus=0        # which GPU to use
-export CUDA_VISIBLE_DEVICES=0
+#export NCCL_DEBUG=INFO
+export NCCL_IB_GID_INDEX=3
+#export GLOG_v=1
+#export GLOG_logtostderr=1
+#export FLAGS_selected_gpus=0        # which GPU to use
+export CUDA_VISIBLE_DEVICES=0,1,2,3
 
 ERNIE_PRETRAINED_MODEL_PATH=./pretrained/
-ERNIE_FINETUNED_MODEL_PATH=./model_finetuned/
+ERNIE_FINETUNED_MODEL_PATH=./ernie_models/step_10
 DATA_PATH=./data/
-
+source /home/work/huangdingbang/.bash_fluid
 # train
 function run_train() {
     echo "training"
-    python run_ernie_sequence_labeling.py \
+    fluid run_ernie_sequence_labeling.py \
+        --mode train \
         --ernie_config_path "${ERNIE_PRETRAINED_MODEL_PATH}/ernie_config.json" \
-        --checkpoints "./checkpoints" \
+        --model_save_dir "./ernie_models" \
         --init_pretraining_params "${ERNIE_PRETRAINED_MODEL_PATH}/params/" \
         --epoch 10 \
-        --save_steps 1000 \
-        --validation_steps 1000 \
-        --lr 2e-4 \
+        --save_steps 5 \
+        --validation_steps 5 \
+        --base_learning_rate 2e-4 \
         --crf_learning_rate 0.2 \
         --init_bound 0.1 \
-        --skip_steps 1 \
+        --print_steps 1 \
         --vocab_path "${ERNIE_PRETRAINED_MODEL_PATH}/vocab.txt" \
         --batch_size 64 \
         --random_seed 0 \
         --num_labels 57 \
         --max_seq_len 128 \
-        --train_set "${DATA_PATH}/train.tsv" \
-        --test_set "${DATA_PATH}/test.tsv" \
+        --train_data "${DATA_PATH}/train2.tsv" \
+        --test_data "${DATA_PATH}/test2.tsv" \
         --label_map_config "./conf/label_map.json" \
         --do_lower_case true \
-        --use_cuda false \
-        --do_train true \
-        --do_test true \
-        --do_infer false
+        --use_cuda true \
+        --cpu_num 4
 }
 
 
 function run_eval() {
     echo "evaluating"
-    python run_ernie_sequence_labeling.py \
+    fluid run_ernie_sequence_labeling.py \
+        --mode eval \
         --ernie_config_path "${ERNIE_PRETRAINED_MODEL_PATH}/ernie_config.json" \
-        --init_pretraining_params "${ERNIE_PRETRAINED_MODEL_PATH}/params/" \
         --init_checkpoint "${ERNIE_FINETUNED_MODEL_PATH}" \
         --init_bound 0.1 \
         --vocab_path "${ERNIE_PRETRAINED_MODEL_PATH}/vocab.txt" \
@@ -52,21 +55,18 @@ function run_eval() {
         --random_seed 0 \
         --num_labels 57 \
         --max_seq_len 128 \
-        --test_set "${DATA_PATH}/test.tsv" \
+        --test_data "${DATA_PATH}/test2.tsv" \
         --label_map_config "./conf/label_map.json" \
         --do_lower_case true \
-        --use_cuda true \
-        --do_train false \
-        --do_test true \
-        --do_infer false
-}
+        --use_cuda true
 
+}
 
 function run_infer() {
     echo "infering"
-    python run_ernie_sequence_labeling.py \
+    fluid run_ernie_sequence_labeling.py \
+        --mode infer \
         --ernie_config_path "${ERNIE_PRETRAINED_MODEL_PATH}/ernie_config.json" \
-        --init_pretraining_params "${ERNIE_PRETRAINED_MODEL_PATH}/params/" \
         --init_checkpoint "${ERNIE_FINETUNED_MODEL_PATH}" \
         --init_bound 0.1 \
         --vocab_path "${ERNIE_PRETRAINED_MODEL_PATH}/vocab.txt" \
@@ -74,14 +74,13 @@ function run_infer() {
         --random_seed 0 \
         --num_labels 57 \
         --max_seq_len 128 \
-        --infer_set "${DATA_PATH}/test.tsv" \
+        --test_data "${DATA_PATH}/test2.tsv" \
         --label_map_config "./conf/label_map.json" \
         --do_lower_case true \
-        --use_cuda true \
-        --do_train false \
-        --do_test false \
-        --do_infer true
+        --use_cuda true
+
 }
+
 
 
 function main() {
